requirements.txt
PyGithub>=2.1.1
sentence-transformers>=2.2.2
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.65.0
watchdog>=3.0.0
chromadb>=0.4.15
pydantic>=2.0.0 